Pop up de eliminar pregunta +15

+++ web/src/pages/OwnerQuestionsPage.js
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react'
 import { connect } from 'react-redux'
+import Swal from 'sweetalert2'
 
 import { fetchOwnerQuestions, deleteQuestion } from '../actions/questionActions'
 import { Question } from '../components/Question'
@@ -16,7 +17,24 @@ const OwnerQuestionsPage = ({ dispatch, loading, questions, hasErrors, redirect,
     }, [redirect, dispatch, userId]);
 
     const onDelete = (id) => {
-        dispatch(deleteQuestion(id))
+        Swal.fire({
+            title: '¿Estas seguro que deseas eliminar esta pregunta?',
+            text: "Será eliminada permanentemente.",
+            icon: 'Advertencia',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sí, eliminar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteQuestion(id))
+                Swal.fire(
+                    'Eliminada',
+                    'Tu pregunta ha sido eliminada.',
+                    'Eliminada con éxito'
+                )
+            }
+        })
     }
